requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python
Pillow
matplotlib
scikit-image
numpy



